Reorganize svg components
