+++ src/components/atoms/svg/Scientech.tsx
